Move getCurrentBiome to biomes.ts
