+++ src/features/island/hud/components/buildings/Buildings.tsx
@@ -21,7 +21,7 @@ import {
   makeUpgradableBuildingKey,
   isBuildingUpgradable,
 } from "features/game/events/landExpansion/upgradeBuilding";
-import { getCurrentBiome } from "features/island/lib/alternateArt";
+import { getCurrentBiome } from "features/island/biomes/biomes";
 
 interface Props {
   onClose: () => void;
