+++ src/features/island/plots/components/Soil.tsx
@@ -4,7 +4,7 @@ import { CropName } from "features/game/types/crops";
 import { PIXEL_SCALE } from "features/game/lib/constants";
 import { CROP_LIFECYCLE, SOIL_IMAGES } from "../lib/plant";
 import { GameState } from "features/game/types/game";
-import { getCurrentBiome } from "features/island/lib/alternateArt";
+import { getCurrentBiome } from "features/island/biomes/biomes";
 
 export type GrowthStage = "seedling" | "halfway" | "almost" | "ready";
