+++ src/features/island/plots/components/TsunamiPlot.tsx
@@ -12,7 +12,7 @@ import tsunamiIcon from "assets/icons/tsunami.webp";
 import { Label } from "components/ui/Label";
 import { secondsToString } from "lib/utils/time";
 import { SOIL_IMAGES } from "../lib/plant";
-import { getCurrentBiome } from "features/island/lib/alternateArt";
+import { getCurrentBiome } from "features/island/biomes/biomes";
 
 const TsunamiPlotComponent: React.FC<{ game: GameState }> = ({ game }) => {
   const [showModal, setShowModal] = useState(false);
